refactor(typescript): convert wokers/global to ts
