+++ package.json
@@ -170,6 +170,7 @@
     "@babel/preset-env": "7.7.4",
     "@babel/preset-typescript": "7.7.4",
     "@types/bunyan": "1.8.6",
+    "@types/cacache": "12.0.1",
     "@types/chai": "4.2.5",
     "@types/convert-hrtime": "2.0.0",
     "@types/fs-extra": "8.0.1",
