+++ test/workers/global/index.spec.ts
@@ -1,24 +1,25 @@
-const globalWorker = require('../../../lib/workers/global');
-const repositoryWorker = require('../../../lib/workers/repository');
-/** @type any */
-const configParser = require('../../../lib/config');
-/** @type any */
-const platform = require('../../../lib/platform');
+import * as globalWorker from '../../../lib/workers/global';
+import * as _repositoryWorker from '../../../lib/workers/repository';
+import * as _configParser from '../../../lib/config';
+import * as _platform from '../../../lib/platform';
+import * as _limits from '../../../lib/workers/global/limits';
 
-jest.mock('../../../lib/platform');
+jest.mock('../../../lib/workers/repository');
 
-const limits = require('../../../lib/workers/global/limits');
+// imports are readonly
+const repositoryWorker = _repositoryWorker;
+const configParser: jest.Mocked<typeof _configParser> = _configParser as never;
+const platform: jest.Mocked<typeof _platform> = _platform as never;
+const limits = _limits;
 
 describe('lib/workers/global', () => {
   beforeEach(() => {
     jest.resetAllMocks();
     configParser.parseConfigs = jest.fn();
-    configParser.getRepositoryConfig = jest.fn();
-    repositoryWorker.renovateRepository = jest.fn();
-    platform.initPlatform.mockImplementation(input => input);
+    platform.initPlatform.mockImplementation(input => Promise.resolve(input));
   });
   it('handles config warnings and errors', async () => {
-    configParser.parseConfigs.mockReturnValueOnce({
+    configParser.parseConfigs.mockResolvedValueOnce({
       repositories: [],
       maintainYarnLock: true,
       foo: 1,
@@ -26,7 +27,7 @@ describe('lib/workers/global', () => {
     await globalWorker.start();
   });
   it('handles zero repos', async () => {
-    configParser.parseConfigs.mockReturnValueOnce({
+    configParser.parseConfigs.mockResolvedValueOnce({
       baseDir: '/tmp/base',
       cacheDir: '/tmp/cache',
       repositories: [],
@@ -34,7 +35,7 @@ describe('lib/workers/global', () => {
     await globalWorker.start();
   });
   it('processes repositories', async () => {
-    configParser.parseConfigs.mockReturnValueOnce({
+    configParser.parseConfigs.mockResolvedValueOnce({
       gitAuthor: 'a@b.com',
       enabled: true,
       repositories: ['a', 'b'],
@@ -54,8 +55,7 @@ describe('lib/workers/global', () => {
 
   it('processes repositories break', async () => {
     limits.getLimitRemaining = jest.fn(() => 0);
-    // limits.getLimitRemaining.mockReturnValueOnce(0);
-    configParser.parseConfigs.mockReturnValueOnce({
+    configParser.parseConfigs.mockResolvedValueOnce({
       gitAuthor: 'a@b.com',
       enabled: true,
       repositories: ['a', 'b'],
@@ -75,7 +75,7 @@ describe('lib/workers/global', () => {
 
   describe('processes platforms', () => {
     it('github', async () => {
-      configParser.parseConfigs.mockReturnValueOnce({
+      configParser.parseConfigs.mockResolvedValueOnce({
         repositories: ['a'],
         platform: 'github',
         endpoint: 'https://github.com/',
@@ -85,7 +85,7 @@ describe('lib/workers/global', () => {
       expect(repositoryWorker.renovateRepository).toHaveBeenCalledTimes(1);
     });
     it('gitlab', async () => {
-      configParser.parseConfigs.mockReturnValueOnce({
+      configParser.parseConfigs.mockResolvedValueOnce({
         repositories: [{ repository: 'a' }],
         platform: 'gitlab',
         endpoint: 'https://my.gitlab.com/',
