+++ test/workers/global/limits.spec.ts
@@ -6,19 +6,19 @@ import {
 
 describe('lib/workers/global/limits', () => {
   describe('init()', () => {
-    it('check defined variables have a value set to zero', async () => {
+    it('check defined variables have a value set to zero', () => {
       const config = { prCommitsPerRunLimit: 3 };
       init(config);
-      const result = await getLimitRemaining('prCommitsPerRunLimit');
+      const result = getLimitRemaining('prCommitsPerRunLimit');
       expect(result).toEqual(3);
     });
   });
   describe('incrementLimit()', () => {
-    it('check increment works as expected', async () => {
+    it('check increment works as expected', () => {
       const config = { prCommitsPerRunLimit: 3 };
       init(config);
       incrementLimit('prCommitsPerRunLimit', 2);
-      const result = await getLimitRemaining('prCommitsPerRunLimit');
+      const result = getLimitRemaining('prCommitsPerRunLimit');
       expect(result).toEqual(1);
     });
   });
